backend/requirements.txt
langchain
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
sentence-transformers
fastapi
uvicorn
chromadb
openai
python-multipart
pypdf
python-dotenv
requests
numpy
langgraph
